separate mongo data from index and make config folder
