+++ server/src/config/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+    PORT: process.env.PORT || 5000,
+    MONGO_DB:{
+        URI: process.env.DB || 'mongodb://localhost/',
+        DB: 'zefaf'
+    } 
+}
